UL-10 Reorganize account_management package
Changed the directory structore of this package. All the resource
management api are moved into a single module views.py. Removed
views package. Removed urls.py moduel as not needed.
