--- project/account_management/views/__init__.py
@@ -1,13 +1,0 @@
-import ninja
-
-
-class ErrorMessageSchema(ninja.Schema):
-    message: str
-
-
-class SupercategoryUnavailableError(Exception):
-    pass
-
-
-class DuplicateValueError(Exception):
-    pass
