+++ project/app/urls.py
@@ -14,10 +14,11 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path, include
+from django.urls import path 
 
+import account_management.views
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("api/", include("account_management.urls")),
+    path("api/", account_management.views.api.urls),
 ]
